Activities Analysis/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
PyYAML>=6.0
python-docx>=1.0.0
